Intro Scene
The exit button is bit scuffed (don't know why), other than that everything appears to be working fine. The shell of it is done. One of the things that needs doing is the background I still haven't settled on a colour palette for that, but I wanted to have something uploaded for this week just in case I forgot about the background.